double->int

+++ src/main/java/com/uci/warehouse/Order.java
@@ -104,12 +104,12 @@ public class Order {
      * @return getXYDistanceMatrix
      * Jindong
      */
-    public double[][][] getXYDistanceMatrix(Map<Integer, double[]> locationMap) {
+    public int[][][] getXYDistanceMatrix(Map<Integer, double[]> locationMap) {
         List<Integer> list = getOrderList();
         //add endpoint at last
         list.add(list.size(),-1);
         locationMap.put(-1, new double[]{0,0});
-        double[][][] m = new double[list.size()][list.size()][2];
+        int[][][] m = new int[list.size()][list.size()][2];
         for (int i = 0; i < list.size(); i++) {
             for (int j = 0; j < list.size(); j++) {
                 if (!locationMap.containsKey(list.get(i)) || !locationMap.containsKey(list.get(j))) try {
@@ -117,8 +117,8 @@ public class Order {
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
-                m[i][j][0] = -locationMap.get(list.get(i))[0] + locationMap.get(list.get(j))[0];
-                m[i][j][1] = -locationMap.get(list.get(i))[1] + locationMap.get(list.get(j))[1];
+                m[i][j][0] = -(int)locationMap.get(list.get(i))[0] + (int)locationMap.get(list.get(j))[0];
+                m[i][j][1] = -(int)locationMap.get(list.get(i))[1] + (int)locationMap.get(list.get(j))[1];
             }
         }
         return m;
